Don't overwrite ExpDate if PayPal returns obfuscated information
There is at least one case of PayPal returning "XXXX" as the ExpDate for an extended period of time in this. API call. If PayPal returns a non-numeric response for the EXPDATE field, Magento will interpret it as Month 0 of the year 2100.

This modification ensures that if such a case occurrs again, Magento will instead retain the expiration month and year it collected during the checkout process instead of replacing it with faulty data.