Improve enabledFlat check when there are a lot of categories
For the menu block the call to `if ($this->flatState->isAvailable()) {` can cost up to 20ms if more then 1k of categories are loaded. This can be optimized by keeping track of the isAvailable return value depending on the used storeId.

A similar approach is done in `vendor/magento/module-catalog/Model/ResourceModel/Product/Collection.php` 

This is a minor improvement, but still worthwhile in my opinion.