Avoid lots of uncessary exceptions during layout XML merge loading
* Introduce a new function (for B/C compat we don't overwrite the protected _loadXmlString) to load and not throw an exception
* Use it in extract handlers -> no need to catch the exception any more

Reason:

When using tools like Sentry, because of there inner workings, such exceptions still might be reported.
While this is not exactly the root cause, it's always cleaner to not throw an exception and immediately catch it, if invalid cases can happen often.

This might even improve performance a bit and improves developer's experience.