ACPT-1413
Removed getProcessConnectionName from lib/internal/Magento/Framework/App/ResourceConnection.php
We cannot have separate connections in this class based on the process id.  The reason is that because some objects have references to the connection object before the fork, and therefore the child would be accessing a different data connection in some classes,  which could break the expected transaction functionality.  In other words, the getProcessConnectionName functionality did not work in cases where objects used by child process had a member variable to the parent process's connection (Mysql Adapter).

We added avoidReusingParentProcessConnection() in Mysql adapter to avoid using the parent's connection.