Fix #37937: IdListBuilder object reuse causes accumulated joins
When async grid update is enabled the instance of IdListBuilder gets reused across multiple grid update tasks. This causes the associated join to the archived grid tables to accumulate for each and results in joining unrelated archived grid tables.

In certain circumstances this results in invoices, shipments and creditmemos incorrectly being excluded from the list of IDs not present in the grid.