ACPT-1599: Fixing DeploymentConfig resets & reloads every time new key that wasn't previously set
Fixing problem ACP2E-2198.
```
[Progress: 861 / 864]
Disabling Maintenance Mode:

In ResourceConnection.php line 162:

[DomainException]
Connection "default" is not defined
```

The root cause of problem is that the `Disabling Maintenance Mode` part of the install uses the first created DeploymentConfig object created in the process, which doesn't get updated, while before that runs, different part of the install that writes to the config is using the second DeploymentConfig object that was created by Laminas ServiceManager.

This commit causes Laminas' ServiceManager to use the same DeploymentConfig object as the first Magento ObjectManager used in the process, which resolves the issue seen in setup:install.