Revert "ACPT-1599: Fixing DeploymentConfig resets & reloads every time new key that wasn't previously set"
This reverts commit a731903ceab6881ca09875eb434a7bc4c5568f9d.

This workaround didn't work for very specific use-case of custom 'etc' path when running setup:upgrade.  This was found in setup-integration tests.  The reason why it didn't work is because the first DeploymentConfig doesn't use the custom paths for 'etc', so it was loading the wrong files.  This problem may still exist for plugins and other objects that use the original ObjectManager's DeploymentConfig instead of the DeploymentConfig object created by Laminas ServiceManager.