Add default empty array in constructor
This class its constructor argument `$tagGenerators` is empty by default in PHP, so that via DI argument types in `di.xml` files the array is filled automatically. This actually means that - without any runtime configuration of this argument - the PHP class has a required argument `$tagGenerators`.

In my cases, this is breaking DI compile when the modules `Magento_DirectoryGraphQl` and `Magento_StoreGraphQl` are disabled (and actually a lot of people will disable this module if they are not building a headless shop). This PR is a simple but required fix.