ACPT-1742: Fixing: Redis Session uses 503.php which uses custom ObjectManager not compatible with ApplicationServer
The 503.php, and its friends, should not be used in ApplicationServer
because Magento\Framework\Error\ProcessorFactory uses its own
ObjectManager instead of the current one.  Therefore, it is not
compatible with ApplicationServer which relies own its own
ObjectManager.

This fix just changes The Redis session wrapper to throw an exception
instead of calling 503.php.

This change will change the behavior when
ConcurrentConnectionsExceededException happens as it will now respond
with 500 instead of 503.