Fix: The Credit Memo Items' base_weee_tax_applied_row_amnt Amount is Incorrect
The setBaseWeeeTaxAppliedRowAmnt() function should be used instead of setBaseWeeeTaxAppliedRowAmount, as the database table field name is base_weee_tax_applied_row_amnt. Due to this error, the value of the sales_creditmemo_item.base_weee_tax_applied_row_amnt database column is always set to the order item value, which is incorrect.