misuse of API … call with `true` deletes this object's data
misuse of API, ... When `true`, the resultant getData(..,true) call shifts the context to the specified field, expecting it to be an array.  The previous Data context is replaced.  In this case, the field is not an array and all of the previous data on this object now returns NULL.  I don't believe loss of this, and all of the other fields was the intended behavior.