Update MetadataConfigTypeProcessor.php
During updates, especially when modules are being removed, some configuration vestiges will cause the `bin/magento setup:upgrade` to fail here, because they are actually missing and cannot be loaded.

This is primarily a convenience to the developer, … with this change  the configuration vestiges are cleaned up once the `bin/magento setup:upgrade` completes.  Without it, the `bin/magento setup:upgrade` fails to start.