check not_is_null before use
During updates, especially when modules are being removed, some configuration vestiges will cause the `bin/magento setup:upgrade` to fail, because they are actually missing and cannot be loaded.

This is primarily a convenience to the developer, … with this change the configuration vestiges, that are cleaned up once the bin/magento setup:upgrade completes are able to run to completion.  Without it, the `bin/magento setup:upgrade` fails to start.