Fix table render of DiInfoCommand
When running the DiInfoCommand under Magento 2.4.7 and PHP 8.3, line 96 mentions to pass the output of `$paramsTable->render()` to the `$output`. However, because of typing issues, this causes a Fatal Error. According the Symfony docs (since a long time) the `render()` method is already giving output, so the output should not be passed to `$output->writeln()` anyway. This PR fixes things.