Stopped marking transaction in integration tests as inactive although still active
Before an integration test is started the testing framework begins a new database transaction so that all changes to the database can be reverted after the test completed. The testing framework also offers a way to easily load data fixtures during this transaction with the help of the  attribute. But in case of faulty fixtures an exception or error can occure. Currently this will prevent the  event from rollback the transaction and so the tests fail with the error message 'Fatal error: Some transactions have not been committed or rolled back in /var/www/html/vendor/magento/framework/DB/Adapter/Pdo/Mysql.php on line 4169'.

The issue is that the catch block in the Transaction event marks the transaction as 'inactive' so that it won't be rolled back at the end of the test, although the transaction is still open. This is fixed now.